% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapPvalue.R
\name{boot2pval}
\alias{boot2pval}
\title{Bootstrap Testing}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{-sample 1}

\item{x2}{-sample 2}

\item{conf.level}{-desired confidence}

\item{iter}{-iterations}

\item{mudiff}{-difference in means of the samples}

\item{test}{-which test to use}
}
\value{

}
\description{
This function will allow us to view the p-value and to determine whether we should reject or fail to reject the null hypothesis
}
\examples{
\dontrun boot2pval(x1=x,x2=y)
}
