% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapIntervalFunction.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{data set from distribution}

\item{fun}{desired function}

\item{alpha}{1-confidence%}

\item{cx}{character expansion factor font size}

\item{...}{}
}
\value{

}
\description{
Bar plot of statistic with point estimate and confidence intervals
}
\examples{
\dontrun myboot2(iter=10000,sam,fun="mean",alpha=.05,cx=1.5,xlab="mean",col="green")
}
