% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval.R
\name{myci}
\alias{myci}
\title{Confidence Interval}
\usage{
myci(x, confint = 0.95)
}
\arguments{
\item{x}{data}

\item{confint}{confidence interval desired}
}
\value{

}
\description{
This function will find a confidence interval for the mean of a given set of data
}
\examples{
\dontrun myci(x,.95)
}
