% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonSample.R
\name{mycltp}
\alias{mycltp}
\title{Poisson Sample}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{lambda}{exponential value in poisson distribution}

\item{...}{}
}
\value{

}
\description{
Poisson Sample
}
\examples{
\dontrun mycltp(n=10,iter=10000,lambda=10)
}
