% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myhyper.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeomtric Distribution}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{is number of iterations}

\item{N}{is the total number of marbles}

\item{r}{number of white marbles}

\item{n}{is the sample size}
}
\value{
puts out a bar graph
}
\description{
Finds bargraph of hypergeometric distribution
}
\examples{
\dontrun myhyper(iter=1000, N=20, r=12, n=5)
}
