% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
mymlnorm
}
\examples{
\dontrun mymlnorm(x=c(10,12,13,15,12,11,10),mu=seq(8,15,length=1000),sig=seq(.1,4,length=1000),lwd=2,labcex=1)
}
