% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyNormalDistribution.R
\name{myncurve}
\alias{myncurve}
\title{Norm Prob Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{is mean}

\item{sigma}{is standard deviation}

\item{a}{is q value}
}
\value{

}
\description{
Will show a shaded normal probability curve
}
\examples{
\dontrun myncurve(mu=10,sigma=5, a=6)
}
